requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
blinker==1.9.0
    # via flask
cachetools==5.5.1
    # via google-auth
certifi==2024.12.14
    # via
    #   -r requirements.in
    #   requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   -r requirements.in
    #   flask
colorama==0.4.6
    # via click
dnspython==2.7.0
    # via pymongo
flask==3.1.0
    # via
    #   -r requirements.in
    #   flask-login
    #   flask-mongoengine-3
flask-login==0.6.3
    # via -r requirements.in
flask-mongoengine-3==1.1.0
    # via -r requirements.in
google-auth==2.38.0
    # via google-auth-oauthlib
google-auth-oauthlib==1.2.1
    # via -r requirements.in
idna==3.10
    # via requests
itsdangerous==2.2.0
    # via flask
jinja2==3.1.5
    # via flask
markupsafe==3.0.2
    # via
    #   -r requirements.in
    #   jinja2
    #   werkzeug
mongoengine==0.29.1
    # via flask-mongoengine-3
numpy==2.2.2
    # via -r requirements.in
oauthlib==3.2.2
    # via
    #   -r requirements.in
    #   requests-oauthlib
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pymongo==4.11
    # via mongoengine
requests==2.32.3
    # via
    #   -r requirements.in
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rsa==4.9
    # via google-auth
urllib3==2.3.0
    # via requests
werkzeug==3.1.3
    # via
    #   -r requirements.in
    #   flask
    #   flask-login
